requirements.txt
Django==4.1
django-debug-toolbar==3.6.0
python-dotenv==0.21.0
sqlparse==0.4.2
psycopg2
djangorestframework
django-request-logging
Pillow
django-filter
drf-yasg
pyjwt